parcial_construccion/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]